addEmployee function

+++ schema.sql
@@ -15,7 +15,7 @@ CREATE TABLE department (
 CREATE TABLE _role (
     id INT AUTO_INCREMENT NOT NULL,
     title VARCHAR(30) NOT NULL,
-    salary DECIMAL,
+    salary INT,
     department_id INT,
     PRIMARY KEY (id),
     FOREIGN KEY(department_id) REFERENCES department(id)
